Crate qdrant_edge

Modules

external

Re-export from external crates used by Qdrant.

internal

Structs

AcornSearchParams

ACORN-related search parameters

BinaryQuantizationConfig

BoolIndexParams

ConditionalInsertOperation

ContextPair

ContextQuery

CountRequest

Count Request Counts the number of points which satisfy the given filter. If filter is not provided, the count of all points in the collection will be returned.

CreateIndex

DateTimeWrapper

Wraps DateTime<Utc> to allow more flexible deserialization

DatetimeIndexParams

DeletePayloadOp

This data structure is used inside shard operations queue and supposed to be written into WAL of individual shard.

DiscoverQuery

EdgeConfig

Full configuration for an edge shard.

EdgeOptimizersConfig

Optimizer-related parameters for the edge shard.

EdgeShard

EdgeSparseVectorParams

User-facing sparse vector parameters.

EdgeVectorParams

User-facing dense vector parameters.

FacetHit

FacetRequest

Facet Request Counts the number of points for each unique value of a payload key.

FacetResponse

FeedbackItem

FeedbackNaiveQuery

Akin to external representation of the query. Unoptimized for scoring.

FieldCondition

All possible payload filtering conditions

Filter

FloatIndexParams

Formula

GeoBoundingBox

Geo filter request

GeoIndexParams

GeoPoint

Geo point payload schema

GeoPolygon

Geo filter request

GeoRadius

Geo filter request

HasIdCondition

ID-based filtering condition

HasVectorCondition

Filter points which have specific vector assigned

HnswIndexConfig

Config of HNSW index

IntegerIndexParams

IsEmptyCondition

Select points with empty payload for a specified field

IsNullCondition

Select points with null payload for a specified field

JsonPath

KeywordIndexParams

MatchAny

Exact match on any of the given values

MatchExcept

Should have at least one value not matching the any given values

MatchPhrase

Full-text phrase match of the string.

MatchText

Full-text match of the strings.

MatchTextAny

Full-text match of at least one token of the string.

MatchValue

Exact match of the given value

MinShould

Mmr

Maximal Marginal Relevance configuration

MultiVectorConfig

NaiveFeedbackStrategy

Trained coefficients for the formula. Specific to a triplet of dataset-smallmodel-bigmodel.

NamedQuery

NamedVectors

Named vectors map.

Nested

Select points with payload for a specified nested field

NestedCondition

OrderBy

Payload

PayloadIndexInfo

Display payload field type & index information

PayloadSelectorExclude

PayloadSelectorInclude

PointIdsList

PointStruct

PointStructPersisted

PointSyncOperation

PointVectorsPersisted

Prefetch

ProductQuantizationConfig

QuantizationSearchParams

Additional parameters of the search

QueryRequest

Internal representation of a universal query request.

Range

Range filter request

RecommendQuery

Record

Point data

ScalarQuantizationConfig

ScoredPoint

Search result

ScrollRequest

Scroll request - paginate over all points which matches given condition

SearchParams

Additional parameters of the search

SearchRequest

DEPRECATED: Search method should be removed and replaced with ShardQueryRequest

SetPayloadOp

This data structure is used inside shard operations queue and supposed to be written into WAL of individual shard.

ShardInfo

SnowballParams

SparseVector

Sparse vector structure

StopwordsSet

TextIndexParams

TypedMultiDenseVector

Type for multi dense vector

UpdateVectorsOp

UuidIndexParams

ValuesCount

Values count filter request

Vector

Single vector.

Vectors

Collection of vectors for a point.

Enums

AnyVariants

BinaryQuantizationEncoding

BinaryQuantizationQueryEncoding

CompressionRatio

Condition

DecayKind

Direction

Distance

Type of internal tags, build from payload Distance function types used to compare vectors

Expression

FacetValue

FieldIndexOperations

Fusion

Language

Match

Match filter request

Modifier

If used, include weight modification, which will be applied to sparse vectors at query time: None - no modification (default) Idf - inverse document frequency, based on statistics of the collection

MultiVectorComparator

OperationError

OrderByInterface

OrderValue

PayloadFieldSchema

PayloadOps

Define operations description for point payloads manipulation

PayloadSchemaParams

Payload type with parameters

PayloadSchemaType

All possible names of payload types

PayloadSelector

Specifies how to treat payload selector

PointId

Type, used for specifying point ID in user interface

PointInsertOperations

PointOperations

QuantizationConfig

QueryEnum

Every kind of vector query that can be performed on segment level.

RangeInterface

Sample

ScalarType

ScoringQuery

Same as Query, but with the resolved vector references.

SnowballLanguage

Languages supported by snowball stemmer.

StartFrom

TokenizerType

UpdateMode

Defines the mode of the upsert operation

UpdateOperation

ValueVariants

VectorInternal

VectorOperations

VectorPersisted

Single vector data, as it is persisted in WAL Unlike [api::rest::Vector], this struct only stores raw vectors, inferenced or resolved. Unlike VectorInternal, is not optimized for search

VectorStorageDatatype

Storage types for vectors

VectorStructInternal

Full vector data per point separator with single and multiple vector modes

VectorStructPersisted

Data structure for point vectors, as it is persisted in WAL

WithPayloadInterface

Options for specifying which payload to include or not

WithVector

Options for specifying which vector to include

Constants

DEFAULT_VECTOR_NAME

Type Aliases

FacetValueHit

OperationResult